LayoutTests/platform/ios-sim-deprecated/accessibility/dom-focus-fires-on-correct-element-expected.txt

link1
heading1

This tests that accessibilityElementDidBecomeFocused will be propagated to the correct parent and cause DOM focus to fire on the correct elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textInsideLink.iphoneLabel is 'AXLabel: link1'
PASS accessibilityController.focusedElement.childAtIndex(0).isEqual(textInsideLink) is true
PASS textInsideHeading.iphoneLabel is 'AXLabel: heading1'
PASS accessibilityController.focusedElement.childAtIndex(0).isEqual(textInsideHeading) is true
PASS successfullyParsed is true

TEST COMPLETE
